Plot of wisp parameters

Function to make nicely formatted violin (or bar) plots of the fitted wisp parameters, including confidence intervals if stat information is available.

Usage

# S3 method for class 'parameters'
plot(
  wisp.results,
  child.lvl = NULL,
  violin = TRUE,
  print.plots = TRUE,
  child.classes = NULL,
  verbose = TRUE
)

Arguments

wisp.results

List, output of the wisp function.

child.lvl

Character string, the child level to be plotted. If NULL, all child levels are plotted.

violin

Logical, if TRUE, plots violin plots for each parameter; if FALSE, uses bar plots.

print.plots

Logical, if TRUE, prints the plots to the console; if FALSE, only returns a list of plots without printing.

child.classes

List, a list of character vectors specifying which child levels to include together in plots. If NULL, all child levels are included in a single plot.

verbose

Logical, if TRUE, prints updates about the plotting process.

Value

List of ggplot objects for parameter plots.